+++ ChangeLog.md
@@ -8,7 +8,7 @@
 ## NEXT
 
 * Be more resilient about missing metadata in CycloneDX SBOMs.
-* `project createbom` uses purl from SW360 if available instead of building it
+* `project createbom` uses source url and purl from SW360 if available
 * `project createbom` adds SW360 project name, version and description to SBOM.
 
 ## 2.0.0 (2023-06-02)